azure_function_ficp_daily/requirements.txt
azure-functions==1.20.0
azure-storage-blob==12.23.0
azure-identity==1.17.1
